ml-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy<2
opencv-python==4.11.0.86
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
kagglehub>=1.0.0
